+++ src/utilities/dom/index.ts
@@ -1,3 +1,4 @@
 export { getClosestFlyoutTarget } from "./flyout";
 export { getShadowRoot } from "./shadowDom";
 export { enableUserSelect, disableUserSelect } from "./userSelect";
+export { findParent } from "./find";